+++ app/models/product-stock/index.ts
@@ -1,0 +1,2 @@
+export * from './product-stock.server'
+export * from './schema'
