+++ app/models/product/product.server.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { ProductSchema, ProductStockSchema } from './schema'
+import { ProductSchema } from './schema'
 import * as AudiophileClient from '~/utils/audiophile-client'
 import getRandom from '~/utils/get-random'
 
@@ -42,11 +42,4 @@ export const homepageProducts = async () => {
     background,
     sideBySide
   }
-}
-
-export const getProductStocks = async(productSlug: string) => {
-  const response = await AudiophileClient.sendRequest('get', `products/${productSlug}/stocks`)
-  const stocks = z.array(ProductStockSchema).parse(response)
-
-  return stocks
 }
