+++ app/models/purchase-cart-item/schema.ts
@@ -1,0 +1,4 @@
+export interface PurchaseCartItemPayload {
+  stock_uuid: string
+  quantity: number
+}
